Add Timestamp to BaseTelegram and clean up TelegramParser
Added a Timestamp property to the BaseTelegram class to store the creation or processing time of the telegram. Updated the TelegramArgs constructor in TelegramParser.cs to set this Timestamp to the current date and time (DateTime.Now) when a new TelegramArgs instance is created. Removed an unnecessary space in the TelegramParser class before the type check for the telegram.